Sistemato il merge dei best-scores e cambiato tipo attributo score_absolute
score_absolute da integer a numeric (perchè viene fatta la media dei voti dei valutatori)

+++ libs/server/best-scores/src/lib/best-competency-score/entities/best-competency-score.entity.ts
@@ -8,7 +8,7 @@ export class BestCompetencyScoreEntity {
   @PrimaryGeneratedColumn()
   id!: number;
 
-  @Column({ type: 'integer', nullable: false })
+  @Column({ type: 'numeric', nullable: false })
   best_score_absolute!: number;
 
   @Column({ type: 'numeric', nullable: false })
